size_t

+++ esphome/components/power_management/power_management_idf.cpp
@@ -165,7 +165,7 @@ void PowerManagement::dump_config() {
 #endif
 }
 
-static const uint8_t PM_BUF_SIZE = 1024;
+static const size_t PM_BUF_SIZE = 1024;
 static char pm_buffer[PM_BUF_SIZE];
 // TBD replace with pm functions when they are available
 // TBD Fix to also work with profile: true
